# api/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "HS-SPORT API",
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/location/states": {
            "post": {
                "tags": [
                    "Location"
                ],
                "summary": "Location states list",
                "description": "Location states list",
                "operationId": "api.location.states",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "seasonId",
                        "in": "query",
                        "description": "season ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/location/city": {
            "post": {
                "tags": [
                    "Location"
                ],
                "summary": "Location current city or default city by most popular school",
                "description": "Location current city or default city by most popular school",
                "operationId": "api.location.city",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "seasonId",
                        "in": "query",
                        "description": "season ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "cityId",
                        "in": "query",
                        "description": "city ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/location/cities": {
            "post": {
                "tags": [
                    "Location"
                ],
                "summary": "Location cities list",
                "description": "Location cities list",
                "operationId": "api.location.cities",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "seasonId",
                        "in": "query",
                        "description": "season ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/location/teams": {
            "post": {
                "tags": [
                    "Location"
                ],
                "summary": "Location teams list",
                "description": "Location teams list",
                "operationId": "api.location.teams",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "seasonId",
                        "in": "query",
                        "description": "season ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "stateId",
                        "in": "query",
                        "description": "state ID",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "cityId",
                        "in": "query",
                        "description": "city ID",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/location/zip-code": {
            "post": {
                "tags": [
                    "Location"
                ],
                "summary": "Location search from zip code",
                "description": "Location search from zip code",
                "operationId": "api.location.zip.code",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "seasonId",
                        "in": "query",
                        "description": "season ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "zipCode",
                        "in": "query",
                        "description": "zip code",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
...
}